Clustering apparently correct; origin mapping is not
We (correctly) deduce that the sample report list maps to a single
group of overlapping reports.  But, our idea of the relative position
of the scanners is wrong.

We get scanner 1 (which overlaps scanner 0) correct, but we get
incorrect results for all the other scanners (which do not overlap
scanner 0).